+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `purge`, `bulkPurge`, and `purgeTrash` to permanantly remove things
+
 ## [14.1.0] - 2026-08-22
 
 ### Added